Always build images with Buildah
Docker and Buildah create images that differ in subtle but meaningful
ways. Docker is able to run Buildah in a container and have it push back
to the Docker daemon, so do that.

As an added benefit, we can choose the image/version of Buildah
regardless of the host operating system.